So I added three things
One was a new recipe for a time machine
The other was a new utterance for a time traveler

The last thing I added was a new Intent. This intent was suppose to
tell the user a command to enter for a cheat. I worked on this third
part for 2.5 hours and in the end it would not work, I don’t know why.
But the purpose is to ask Alexa for,”what is the Atlantis mode
command”, and then Alexa would tell you. But as I mentioned I could not
get this to work.