BART Sessions
I’ve gotten the program to remember the user’s home, but whenever I try
to ask for the trains from the home (or explicitly name a station), I
get an error saying that the BART system is down. I’m not sure if this
is true, or if it’s a problem on my end, but I’ll come back to it later
and see if it works.